# manifest: improve Annotator interface with generics