--- a/internal/manifest/level_metadata.go
+++ b/internal/manifest/level_metadata.go
@@ -140,31 +140,6 @@ func (lm *LevelMetadata) Find(cmp base.Compare, m *FileMetadata) LevelSlice {
 	return LevelSlice{}
 }
 
-// Annotation lazily calculates and returns the annotation defined by
-// Annotator. The Annotator is used as the key for pre-calculated
-// values, so equal Annotators must be used to avoid duplicate computations
-// and cached annotations. Annotation must not be called concurrently, and in
-// practice this is achieved by requiring callers to hold DB.mu.
-func (lm *LevelMetadata) Annotation(annotator Annotator) interface{} {
-	if lm.Empty() {
-		return annotator.Zero(nil)
-	}
-	v, _ := lm.tree.root.Annotation(annotator)
-	return v
-}
-
-// InvalidateAnnotation clears any cached annotations defined by Annotator. The
-// Annotator is used as the key for pre-calculated values, so equal Annotators
-// must be used to clear the appropriate cached annotation. InvalidateAnnotation
-// must not be called concurrently, and in practice this is achieved by
-// requiring callers to hold DB.mu.
-func (lm *LevelMetadata) InvalidateAnnotation(annotator Annotator) {
-	if lm.Empty() {
-		return
-	}
-	lm.tree.root.InvalidateAnnotation(annotator)
-}
-
 // LevelFile holds a file's metadata along with its position
 // within a level of the LSM.
 type LevelFile struct {